CollectiveAxiom

Documentation

checkAllEqual :: (Eq a, Ord a, Show a, Num a) => [HChoice3] -> HChoice3 -> Bool

only a test function

checkCollectiveRationality :: (Eq a, Ord a, Show a, Num a) => [HChoice3] -> Bool

makeGraphs :: (Eq a, Num a) => [HChoice3] -> (BinRel, BinRel, BinRel, BinRel)

makes graphs representing R0's for each member | e.g. (h0p1 = -1 = h0p1 = 1) = h1 R h0

allElements :: BinRel -> Bool

anyElement :: BinRel -> Bool

restrictionsFromSum :: BinRel -> [Int]

Returns indices columns of the preference relation matrix which puts a restriction on the individuals

takeCols :: [Int] -> BinRel -> BinRel

checkI :: BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> Bool

checkIII :: BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> Bool

checkIV :: BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> Bool

Logic: remove all rows for which there are no restrictions from sum (if zero rows remain - immediately conclude rational: null js)

checkV :: BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> Bool

logic: if in a any row of the or-ed pref rels, all elements are true than this is violated

checkVI :: BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> BinRel -> Bool

checkIndividualRationality :: (Eq a, Ord a, Show a, Num a) => HChoice3 -> Bool

checkPathIndividualRationality :: (Eq a, Ord a, Show a, Num a) => [HChoice3] -> Bool

this was for the household data