Encode

Documentation

checkConsumption :: [Double] -> [Double] -> Double

Check if consumption x is affordable under price regime p

encodeConsumption :: [[Double]] -> [[Double]] -> [[Double]]

Constructs choice path of one individual: periods (outer), encoded polyhedra (inner) | inner part refers to feasibility (checkConsumption) of choice in period t (outer) | when budget set would be s (inner) | rows == cols and zero diagonal by construction

encodeGroupConsumption :: [[Double]] -> [[[Double]]] -> [[[Double]]]

Takes a vector of vectors of consumptions representing each member in the group (outer) | and constructions a path of encoded group consumptions

choicesPerPeriod :: Int -> [[[IChoice]]]

Returns all possible combinations of choices on in | Level 0 (outermost): periods | Level 1: choice set for household in respective period | Level 2: particular choice of household | Level 3 (innermost): choice per individual (polyhedra)

combinations :: Int -> [[[IChoice]]]

combinationsPeriod :: Int -> Int -> [[IChoice]]

choicesPerComposition :: Int -> [[[IChoice]]]

combinePeriods :: [[a]] -> [[a]]

Takes list of period choices and constructs all resulting choice combinations | [[]] is unit-element of operation * on list applicative functor | Each element is one choice path | depr

makeDecisionForest :: ([HChoice3] -> Bool) -> [HChoice3] -> Bool -> [[[IChoice]]] -> Forest RChoice

Constructs a forest from decision space and crawles the branches to check rationality | once an irrational branch is found, the rationality of all child branches is set to false | one-elements (start with): makeDecisionForest [] True choicespace

forestToChoicespace :: Forest RChoice -> [RChoice]

paths :: Tree a -> [[a]]

classifyPaths :: Forest (HChoice3, Bool, Int) -> Paths

asAssocMatrix :: Forest (HChoice3, Bool, Int) -> [Int] -> AssocMatrix

trAssocMatrix :: AssocMatrix -> AssocMatrix

asAssocMatrix' :: Int -> [Int] -> [[Int]] -> AssocMatrix

Histogram uses a little (dirty) trick to also count empty categories: first add one of each index, then decrease all length by one asSparseVector :: [[HChoice3]] -> [[RChoice]] -> GMatrix asSparseVector a b = mkSparse $ asAssocVector a b

depr asAssocVector :: [[HChoice3]] -> [[RChoice]] -> AssocMatrix asAssocVector sampledPaths choicesPerPeriod = histogram where histogram = [((head l, 0), fromIntegral (length l - 1) / fromIntegral (length sampledPaths)) | l <- group (sort $ indices ++ [0..(nChoicesPerPeriod*periods-1)])] nChoicesPerPeriod = (head . map length) choicesPerPeriod periods = length choicesPerPeriod indices = join $ map (zipWith (offset x -> offset x) (offsetters nChoicesPerPeriod)) indices' indices' = fromJust $ mapM (sequence . (sp -> zipWith (curry lookForIndex) sp choicesPerPeriod)) sampledPaths

|depr combinations :: Int -> [[[[Double]]]] combinations p = map (t -> household t ++ members t) [0..(p-1)] -- not replicate3, first one without zero then append where household t = replicateM 2 $ map (squeeze (-1.0) t) powerset ++ map (squeeze 1.0 t) powerset members t = replicateM 2 (map (squeeze 0.0 t) powerset) squeeze d t c = take t c ++ [d] ++ drop t c powerset = replicateM (p-1) [-1.0, 1.0]

offsetters :: [Int] -> [Int -> Int]

getChoicesPerPeriod :: [Tree RChoice] -> [[RChoice]]

lookForIndex :: (HChoice3, [RChoice]) -> Maybe Int