Simulation

Documentation

classifiedPaths :: [([HChoice3], [Int], (Bool, Bool))]

collectiveFilter :: ([HChoice3], [Int], (Bool, Bool)) -> Bool

noncoorpFilter :: ([HChoice3], [Int], (Bool, Bool)) -> Bool

irrationalFilter :: ([HChoice3], [Int], (Bool, Bool)) -> Bool

weakCollectiveFilter :: [[HChoice3]] -> ([HChoice3], [Int], (Bool, Bool)) -> Bool

Paths that are consistent with the aggregate restrictions based on Cherchye et al, not using singles | Only heads are considered, this should be done with [path, [], []] for non-path couple observations

possibleFromAggregate :: [[HChoice3]] -> [HChoice3]

rationalPathsAndIndices :: ([[HChoice3]], [[Int]])

generateChoices :: PrimMonad m => Int -> Int -> Double -> Maybe (Int, [Double]) -> Gen (PrimState m) -> m (SparseMatrix Double, [[[HChoice3]]])

worstCases :: [[HChoice3]] -> [[[HChoice3]]] -> [[[HChoice3]]]

pickleResults :: FilePath -> [((Int, Double, Maybe (Int, [Double])), ([Double], [Double]))] -> IO ()

unpickleResults :: FilePath -> IO [((Int, Double, Maybe (Int, [Double])), ([Double], [Double]))]

simulationPerPN :: Int -> Int -> Int -> Double -> Maybe (Int, [Double]) -> Maybe Double -> IO ([Double], [Double])

resumeSimulation :: Int -> Int -> FilePath -> (Int -> Int) -> Maybe (Int, [Double]) -> Maybe Double -> Int -> Double -> IO ([Double], [Double])

plotPower :: String -> FilePath -> Int -> IO ()

plotSize :: String -> FilePath -> Int -> IO ()